requirements.txt
attrs==20.2.0
black==20.8b1
cattrs==1.0.0
flake8==3.8.1
ipdb==0.13.4
isort==5.2.0
mypy==0.782
requests==2.24.0
pyperclip==1.8.1
pytest==6.1.1
